Move AlState.
